requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asyncpg==0.29.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.112.1
fastapi-swagger==0.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.7
kombu==5.4.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
pydantic==2.8.2
pydantic_core==2.20.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.9
redis==5.0.8
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
